move date encoding and decoding extensions to DateCoding
